Fix StellenT overflow for deep term structures
Change StellenT type from unsigned short (max 65535) to unsigned int
to prevent overflow errors when processing problems with deeply nested
terms that exceed the path stack limit.

+++ sources/NF/NFBildung.c
@@ -33,8 +33,8 @@
 #define BOB_NFCallback(x,y,z) /* nix tun */
 #endif
 
-typedef unsigned short StellenT;
-#define MaxStelle USHRT_MAX
+typedef unsigned int StellenT;
+#define MaxStelle UINT_MAX
 
 #define VERBOSE 1
 extern BOOLEAN drucken;
